lessons/200-how-to-set-up-argocd-for-production-eks-ecr-helm-app-of-apps-image-updater-gitops/k8s-private-repo/requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
idna==3.7
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
